Add C extension engine

+++ motep/potentials/mtp/__init__.py
@@ -3,5 +3,5 @@ import numpy.typing as npt
 from ase import Atoms
 
 
-def get_types(atoms: Atoms, species: list[int]) -> npt.NDArray[np.int64]:
-    return np.fromiter((species.index(_) for _ in atoms.numbers), dtype=int)
+def get_types(atoms: Atoms, species: list[int]) -> npt.NDArray[np.int32]:
+    return np.fromiter((species.index(_) for _ in atoms.numbers), dtype=np.int32)
